move PathMatcher to interactive
